+++ firmware/shared/src/io/Io_SharedStackWaterMark.h
@@ -12,7 +12,7 @@ struct stack_watermark
     uint32_t stack_size;
     float    watermark_threshold;
     void (*log_error)(uint8_t);
-    void (*stack_remaining)(uint8_t);
+    void (*stack_remaining)(float);
 };
 
 /**
